Default affiliation and entitlement to []
AD syncinator and I'm sure others always assumes that the affiliations
and entitlements are at least blank arrays, the API doesn't seem to
enforce it one way or the other and you cannot send blank arrays through
it. So to correct the issue, default affilations and entitlements to a
blank array.